[APT-1178] Publish our Vulnerability Disclosure Policy
this introduces a brand new security page which lists our security policies
(currently just the Vulnerability Disclosure Policy) and some additional
information about our stance on security, reporting a vulnerability, and
how to contact us with concerns. It also adds a Security link to the footer
and consolidates the links to our legal docs.